Add reusable target .prompt-yesno.
A reusable target to prompt the user to make a choice.
Usage:
      make .prompt-yesno message="Is it yes or no?" 2> /dev/null

The make target will exit on an error if anything other than y or Y was entered from stdin.
We redirect the make stderr with 2> /dev/null to not echo an error message.
The user will be prompted with:
    $ Is it yes or no? (Y/N) :

To use in a target you would typically put it in an if statement like such:

uncertain-target:
	@if make .prompt-yesno message="Do you want to abort?" 2> /dev/null then \
	    exit; \
	fi
	@echo "Ok you asked for it..."

The echo command will only be called if user enters a y or Y when prompted with

$ Do you want to abort? (Y/N) :