New addition PHPUnit Test Helpers.
It wouldn't quite be the Foundation Makefile if it doesn't get something new to the pot, now would it.
Learnt my lesson with the issues so consider this one tested and closed.
I love the way the info works so well for the check if installed, grep is so handy for a quiet fail, or display what you were after and don't react.
The pear/peckl info has so much uhm info that you need to trim it down, I might very likely start ronverting the others too. See hawe the install doesn't have a .check-foundation it's because it get's all that from info in any case. 3 targets all rolled into one oops two. =)