Commented out oddities.
Not sure where this is heading but it lost me along the way.
I think the first item would rather be saying something like:

@-rm -f

But then the last one is really odd, besides the fact that if curl failed it will never be called because make aborts there, it is now attempting to run the file we just deleted.
Will have to look at this again... commented now to avoid syntax faults.